tokio: Reference count streams
If two different tasks create a stream from the same procedure call, the
server will assign the same ID to both. When the first of the two
streams is dropped it's removed from the server, breaking the other
task's stream.

This adds reference counting to streams, only removing the stream on
drop if no other tasks are using it.